code separated in modules add tests for getAllTasks in TaskCalendar
