+++ .eslintignore
@@ -3,3 +3,4 @@
 /dist
 /prod
 Readme.md
+.vscode
