+++ .github/workflows/deploy-gh-pages.yml
@@ -21,4 +21,4 @@ jobs:
       - name: Deploy 🚀
         uses: JamesIves/github-pages-deploy-action@v4
         with:
-          folder: dist # The folder the action should deploy.
+          folder: dist
