+++ .lintstagedrc
@@ -1,0 +1,3 @@
+{
+  "*.{ts,mts,cts,tsx,js,mjs,cjs,jsx,json,md,css}": ["prettier --write"]
+}
