+++ jest.config.ts
@@ -8,7 +8,7 @@ import type { Config } from "jest";
 const config: Config = {
   // All imported modules in your tests should be mocked automatically
   // automock: false,
-
+  preset: "ts-jest",
   // Stop running tests after `n` failures
   // bail: 0,
