+++ src/index.ts
@@ -1,4 +1,12 @@
 import "./style/style.css";
 import { TaskCalendar } from "../src/TaskCalendar";
+import { FirebaseTaskService } from "../src/FirebaseTaskService";
+import { LocalStorageTaskService } from "../src/LocalStorageTaskService";
+import { TaskCalendarUI } from "./TaskCalendarUI";
+
+const localTaskService = new LocalStorageTaskService("local_tasks");
+const firebaseTaskService = new FirebaseTaskService();
+
+const taskCalendar = new TaskCalendar(localTaskService, firebaseTaskService);
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-const taskCalendarApi = new TaskCalendar("myNamespace");
+const taskCalendarUI = new TaskCalendarUI(taskCalendar);
