+++ webpack.config.js
@@ -1,6 +1,5 @@
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
-const webpack = require("webpack");
 const path = require("path");
 
 module.exports = {
@@ -49,6 +48,5 @@ module.exports = {
       filename: "index.html",
     }),
     new CleanWebpackPlugin(),
-    new webpack.HotModuleReplacementPlugin(),
   ],
 };
